GH-68: Derived Attribute Issue Preventing Proper Setting Of OSSEC Server IP
Replace all references to `node['ossec']['agent_server_ip']` with `node['ossec']['conf']['agent']['client']['server-ip']` so as to avoid the derived attribute issue.  I acknowledge that this creates a breaking change for folks who rely on the old attribute so waiting for feedback before submitting this fix upstream. It appears to work for my purposes though.